+++ client/attestation/verification_test.go
@@ -32,7 +32,7 @@ func TestResponse(t *testing.T) {
 		abi          string
 		mic          string
 		lut          uint64
-		round        uint32
+		round        uint64
 		hash         string
 	}{
 		{
